pyproject.toml
[project]
name = "video-clone-backend"
version = "0.1.0"
description = "FastAPI backend for video clone application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioboto3>=15.5.0",
    "alembic>=1.18.1",
    "asyncpg>=0.30.0",
    "email-validator>=2.0.0",
    "fastapi>=0.128.0",
    "firebase-admin>=7.1.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sentry-sdk[fastapi]>=2.0.0",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.40.0",
]

[project.scripts]
dev = "uvicorn main:app --reload --host 0.0.0.0 --port 8000"
